Added rotor gyroscopic damping matrix, some CT-Opt documentation:
B_gyro now contains a matrix for the lienarized gyroscopic effects of
each rotor, as represented by a damping matrix that creates cross
couplings in the rotational degrees of freedom.

Fixed a couple bugs related to tower base and turbine outputs in the
results dict.

In the FOCTT example YAML, corrected outdated aeroMod to aeroServoMod
so that rotor response is included.

Added a new documentation page for MHK capabilities being added for
CT-Opt.