Function to compute dynamic tension on mooring lines
This is based on an algebraic approximation by Aranha and Pinto, "Dynamic tension in risers and mooring lines: an algebraic approximation for harmonic excitation".

This version is still for comparisons with MoorDyn, so there is a part of the code that is used to read a MD output file and use the motions from there. This will be removed after the validation is done.