Workaround for diffraction loads on QTFs computed with RAFT
The slender-body approximation adopted in RAFT to compute QTFs is too restrictive in terms of wave length for some hull geometries (e.g., the slow surge motion of the OC4-DeepCwind and the UMaine VolturnUS-S semisubmersibles). This commit introduces a workaround that is based on the analytical solution for the second-order diffraction wave loads on a bottom-mounted surface-piercing vertical circular cylinder.

This correction is in the same spirit as the MacCamy-Fuch correction for first-order wave loads but applied to second-order wave loads.