Moved calculation of QTFs to compute them only once
Before this commit, RAFT computed different QTFs for each sea state. Even though this might be strictly more correct, this is a somewhat gray area. Theoretically, the QTFs should not depend on the sea state, as first-order motions are linear, but this gets tricky once we include the linearized drag.

In any case, the differences for each sea state are not worth the cost of recomputing the QTFs. Most software, such as OpenFAST, OrcaFlex and so on just use the same QTFs for all their analyses.

Hence,  this commit moves the calculation of the QTFs to be before the solution of system dynamics. The RAOs are computed using a white-noise wave with unit amplitude. To avoid code overlap with solveDynamics (within raft_model.py), this commit also moves part of that code to a function called calcMotions within raft_fowt.py