Bug fix for members with more than two stations
This part of the code was supposed to generate nodes progressively for each strip along the member, from the bottom station until the top one.

However, we were creating a node corresponding to the end station after each strip, resulting in multiple instances of the end node at incorrect positions in the arrays. This affected only members with more than two stations, for instance the offset columns of the OC4-DeepCwind platform.

This commit fixes this bug by moving the creation of the end node outside the for loop.