Fix drag formulation for circular cylinders
The drag force transverse to a circular element should be parallel to the relative transverse fluid velocity, but this was not the case with the previous code. This commit fixes that.

This modification does not affect rectangular cylinders. For rectangular cylinders, we still treat the flow along each direction independently.