IMPORTANT: Change in overhang convention is just a warning now
Because some people are using RAFT to analyze downwind turbines, we do NOT change the sign of overhang anymore.

To analyze upwind turbines, the overhang needs to be negative (same convention as OpenFAST). If the overhang is positive, RAFT issues a warning to let the user know that we changed overhang convention from previous versions, but treats the turbine as downwind.

We will probably remove this warning after a few releases.