Removed a lot of commented pieces of code
Code clean up. I tried to keep some chunks of commented code that seemed to be work in progress. Removed only commented code that didn't seem to be used anymore. I didn't clean parts of the code that I don't know well. If we need, we can retrieve the removed comments using git.