Bug fix in helper functin getRMS()
The calculation of the RMS value was missing a factor of 0.5 (within the sqrt). This bug fix impacts the calculation of the linearized drag force/damping, so the affected tests were updated accordingly. It also impacts the standard-deviation, maximum, and minimum values that were printed to the console.

This wasn't a problem in the original code because the calculation of the wave amplitudes also missed that factor of 0.5, so they cancelled out. But I forgot to add this factor in getRMS() when we fixed the calculation of the wave amplitude.