Update true values of the shared mooring test due to using subsystem
Due to the lumped mass capability (which is not currently tested), we now convert mooring lines to subsystem before computing tensions. That changes the size of  the arrays stored in model.results['case_metrics'][iCase]['array_mooring'][metric] used in test_analyzeCases(), as the 7 lines of the test case become 5 due to aggregating composite lines into a single subsystem.