Bugfix to account for rotor yaw control in solve statics
The rotor orientation is set internally depending on yaw_mode, but this was done only after computing the inertial properties of the system in solveStatics(). Consequently, the weight of the RNA was acting as if the turbine was at its initial position regardless of yaw_mode.
This commit fixes that, and now rotor orientation is updated before computing the inertial properties of the system. I'm still not sure if this fix is 100% correct, so this modification needs to be revised before merging rc1.3.2 into main.

The affected tests are the ones with wind in test_model.py. The corresponding true values were updated.