Added capability for lumped inertia properties in the platform
The user can now specify points with inertial properties (mass and moments of inertia) that are added to the total inertia of the floating platform. This is particularly useful for cases where we only have the total mass and moments of inertia of the platform.

Added the relevant section (commented out) in `examples/VolturnUS-S _example.yaml` to exemplify the usage of this new capability.

New test case `VolturnUS-S-pointInertia.yaml`, which is run in`test_fowt.py` and in `test_model.py`