Included added mass from BEM to solveEigen()
model.solveEigen() now includes added mass from BEM (wave radiation) when solving for the eigenfrequencies and eigenmodes of the system.

We are considering the added mass for frequency w=0. It would be more correct to use some sort of iterative approach to consider the added mass at the actual frequency of motion, but this is tricky because we are solving for the coupled problem.

We might get back to this in the future, but using the added mass for w=0 should be good enough in most cases.