Added Node class
The Node class will be used to describe the motions of the structure in the structural/multibody formulation.

A way to see the new structural model:
We could say that RAFT currently solves for the motions of a single node with 6dofs, as all the inertia, stiffness, and whatnot is lumped at the platform reference point. Our plan is that RAFT will solve for the 6dofs of an arbitrary number of nodes, which could be part of a rigid or flexible member.

The Node class is not used by anything in RAFT yet.