Changes to FOWT class to include nodes, joints, more dofs, and whatnot
Major changes to the FOWT class to include nodes, joints, flexible members,  a variable number of dofs (which was assumed 6), etc.

That includes new variables to deal with the full set of dofs and the reduced set of dofs, and methods to create joints, attach members to joints, and to evaluate the reduced degrees of freedom. An important method is fowt.computeTransformationMatrix(), which evaluates the matrix that transforms from the set of reduced degrees of freedom to the set of full degrees of freedom.