Included a structural Node in the Rotor class
The Rotor class now has a Node at the RNA Reference Point, which is the point at the top of the tower around which the nacelle yaws.

This Node object is responsible for
- providing the position and orientation of the RNA Reference Point in the global coordinate system. This is the same job as fowt.r6 used to do, but now allowing for relative motions between the tower and the FOWT. This relative motion is usually due to flexibility, in particular torsion and bending.

- Store the inertia matrix of the RNA in the global reference frame.

- Store the loads on the RNA, transmitting those loads to the rest of the structure via the joint between the tower and the RNA.

Once flexibility/multibody capabilities are complete, the motion of this Mode will be determined by solving the equations of motion of the FOWT.  The position of the Rotor's Node has to be set using fowt.setNodesPosition(), which is called within fowt.setPosition().

This Node does not follow turbine yaw, as it is set by an independent procedure. Yaw control is NOT modified by this commit.