Minor bug fix in getInertia() for zero-mass members
For zero-mass members, we were returning a center of mass equal to np.array([0,0,0]), which is an array of integers. Now returning np.zeros(3), which is an array of floats.

This does not affect the code we have in master, but affects new code that we are working on. Could potentially affect code under development by other people, so pushing this here to make sure that this gets integrated into all branches as soon as v1.4.1 is released.