New input section 'additional_effects' for additional loads and inertia
The user can now provide an additional 6-dof mean load acting on the platform.

The previous option to specify lumped inertia under 'pointInertias' was renamed and is now within the input section 'additional_effects'.

The additional effects are identified by a 'type' key. For now, only 'mean_load' and 'point_inertia' are supported, but planning to include stiffness and damping in the future