Modified fowt.calcStatics to work with multiple bodies
The stiffness matrices and mean load vectors due to hydrostatics and weight are now computed in the reduced degrees of freedom. This is done by computing them in the full degrees of freedom (i.e., individually for each member/component) and assembling the reduced matrices/vectors with the transformation matrix. For the stiffness matrix, we also use the derivative of the transformation matrix to account for the 'geometric stiffness' term (not sure if this is the proper term).