Compute total mass of the FOWT for an arbitrary number of dofs
Compute the total mass from self.M_struc, which has dimension (nDOF, nDOF), using the same idea used in the helper function getMassAndCenterOfBeam(). Cannot use the same function because getMassAndCenterOfBeam() assumes that the input matrix is (6Nnodes, 6Nnodes), i.e. that it has all translations and rotations, but the reduced set of dof doesn't have all dofs.

Also updated getMassAndCenterOfBeam() to include an element-wise multiplication that is needed in fowt.calcStatics() to remove couplings with other dofs. This is not actually needed in getMassAndCenterOfBeam(), as the inertia submatrices of the beam are defined wrt the nodes themselves, but it doesn't hurt and it helps to keep consistency.