Tower base loads for flexible towers
For flexible towers, the tower base loads are now computed based on the finite-element stiffness matrix of the member. For rigid towers, the procedure is the same as before.

In the future, we could compute the loads on the joints and keep a single workflow for computing the tower based loads of both types of tower.