Workaround/bugfix in model.adjustBallast()
model.adjustBallast() uses member headings to exploit floater symmetry but this information was no longer stored at member level (see https://github.com/WISDEM/RAFT/issues/82).
This commit provides a workaround to that problem. In the future, we should overhaul model.adjustBallast() to make it more general, and probably move it to the FOWT class now that we can have multiple FOWTs in a RAFT model