Modified freq discretization of some tests to run faster
Some tests had a very fine frequency discretization that only makes sense for real analyses. Reduced that discretization to run tests faster.

Also added frequency vector to the pickle files with the true values to make it easier to plot the response, which is helpful to debug or check new true values when something changes. Needed that to compare results with the new frequency discretization against the results with the previous discretization.

Also the plots were using angular frequencies but xlabel was in Hz. The plots are properly in Hz now.