Added optional flag to plot the fowt with axes around it
Before this commit, RAFT fowt.plots() would zoom in the FOWT if not plotting the mooring system, which is specified by plot_ms.
Now, we have an optional flag axes_around_fowt that overwrites that behaviour and lets you choose whether you want to zoom in or not. It works like this:

- If the flag is not provided, fowt.plot() just like before: it zooms in the FOWT if not plotting the mooring system, it does not zoom in if plotting the mooring system.
- If axes_around_fowt==True: Zoom in the FOWT regardless of plotting the mooring system or not.
- If axes_around_fowt==False: Zoom in the FOWT regardless of plotting the mooring system or not. In that case, the function uses the axes limits of the ax object passed to it. Users can use MoorPy's helper function `set_axes_equal` to ensure that axes have the same aspect ratio, just like done in model.plot()